1- Transfer Learning/requirements.txt
keras==3.6.0
Keras-Preprocessing==1.1.2
numpy==2.0.2
requests==2.32.3
tensorflow==2.18.0
pillow~=11.1.0
